requirements.txt
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
websocket-client>=1.6.0
python-dotenv>=1.0.0
requests>=2.31.0
scipy>=1.11.0
aiohttp>=3.9.0
schedule>=1.2.0
colorlog>=6.7.0
